Add Danger CI
Summary:

OK this is a bit complicated. As of now this will work on branch-based PRs
all the time and will often fail on fork-based PRs (but not always). SOOO, what
we do is DETECT that failure and *NOT* fail the PR, but if there are real
issues detected, then we do fail the PR.

There's basically two problems going on here:

1. The GITHUB token given to PRs cannot comment on the PR, and while we
could make a token that does, it would have to have `public_repo` which gives
it full read/write access to the all public repos. Boo. Details throughout
https://github.com/danger/danger/issues/1103

2. There's some combination of bugs/race conditions that cause Danger to
not always see the PR commit, even though it has both remotes. This is discussed
here:
https://github.com/danger/danger/issues/1103#issuecomment-610055193

This works around and gives consistent feedback for internal users and
sometimes-feedback for external users. It's a step in the right direction.

Test Plan:

Signed-off-by: Phil Dibowitz <phil@ipom.com>